+++ projectile.el
@@ -80,6 +80,35 @@ attention to case differences."
         (and (>= start-pos 0)
              (eq t (compare-strings suffix nil nil
                                     string start-pos nil ignore-case))))))
+
+  ;; Improved (no more stack overflows) in Emacs 24.5
+  (eval-after-load 'etags
+    '(when (< emacs-major-version 25)
+       (defvar etags--table-line-limit 500)
+       (defun etags-tags-completion-table ()
+         (let (table
+               (progress-reporter
+                (make-progress-reporter
+                 (format "Making tags completion table for %s..." buffer-file-name)
+                 (point-min) (point-max))))
+           (save-excursion
+             (goto-char (point-min))
+             (while (not (eobp))
+               (if (not (re-search-forward
+                         "[\f\t\n\r()=,; ]?\177\\\(?:\\([^\n\001]+\\)\001\\)?"
+                         (+ (point) etags--table-line-limit) t))
+                   (forward-line 1)
+                 (push (prog1 (if (match-beginning 1)
+                                  (buffer-substring (match-beginning 1) (match-end 1))
+                                (goto-char (match-beginning 0))
+                                (skip-chars-backward "^\f\t\n\r()=,; ")
+                                (prog1
+                                    (buffer-substring (point) (match-beginning 0))
+                                  (goto-char (match-end 0))))
+                         (progress-reporter-update progress-reporter (point)))
+                       table))))
+           table))))
+
   )
 
 (defun projectile-trim-string (string)