+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 ### Bugs fixed
 
 * [#871](https://github.com/bbatsov/projectile/issues/871): Stop advice for `compilation-find-file` to override other advices.
+* [#557](https://github.com/bbatsov/projectile/issues/557): stack overflow in `projectile-find-tag'.
 
 ## 0.13.0 (10/21/2015)